Extract media_scanner into separate module
