+++ src/media_scanner/mod.rs
@@ -1,0 +1,4 @@
+mod routes;
+mod utils;
+
+pub use routes::init_routes;
